# docs/api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GOF API",
        "description": "API REST do GOF (Gestão Operacional de Frotas). Todas as respostas, inclusive erros, usam o envelope {status, message, errors, data[, meta]}.",
        "license": {
            "name": "Proprietary"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1 (mesma origem, via nginx)"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Lista usuários",
                "description": "Permissão: users.view (A). Sem escopo de filial. Busca `q` em name, username e email. Ordenação: name, username, email, role, created_at, last_login_at (`-` = desc; padrão name).",
                "operationId": "usersIndex",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Page"
                    },
                    {
                        "$ref": "#/components/parameters/PerPage"
                    },
                    {
                        "$ref": "#/components/parameters/Search"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "-last_login_at,name"
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "operator",
                                "mechanic",
                                "leader",
                                "admin"
                            ]
                        }
                    },
                    {
                        "name": "branch_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/IsActive"
                    },
                    {
                        "$ref": "#/components/parameters/WithTrashed"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Envelope"
                                        },
                                        {
                                            "required": [
                                                "meta"
                                            ],
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/User"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Cria usuário",
                "description": "Permissão: users.manage. E-mail duplicado entre usuários ativos → 422; perfil não-admin sem `branch_id` → 422; filial inexistente ou excluída → 422.",
                "operationId": "usersStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserEnvelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/users/{user}/restore": {
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Restaura usuário excluído",
                "description": "Permissão: users.manage. 409 se não estiver excluído, se o e-mail ou o username já estiver em uso por um usuário ativo ou se a filial vinculada estiver excluída. Os tokens revogados na exclusão não voltam.",
                "operationId": "usersRestore",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restaurado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserEnvelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "409": {
                        "$ref": "#/components/responses/Conflict"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "required": [
                    "id",
                    "name",
                    "username",
                    "email",
                    "role",
                    "branch",
                    "is_active",
                    "last_login_at",
                    "created_at",
                    "updated_at",
                    "deleted_at"
                ],
                "properties": {
                    "id": {
                        "description": "Usuário no CRUD /users (contrato v1.3). Nunca inclui `password` nem `remember_token`.\nO usuário de /auth/* usa UserResource (schema AuthUser).",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Ana Souza",
                        "maxLength": 120
                    },
                    "username": {
                        "type": "string",
                        "pattern": "^[a-z0-9]{3,30}$",
                        "example": "anasouza"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "ana@example.com",
                        "maxLength": 190
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "operator",
                            "mechanic",
                            "leader",
                            "admin"
                        ]
                    },
                    "branch": {
                        "description": "Filial `{id, code, name}`; null = todas (somente admin).",
                        "required": [
                            "id",
                            "code",
                            "name"
                        ],
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "code": {
                                "type": "string",
                                "example": "FIL-001"
                            },
                            "name": {
                                "type": "string",
                                "example": "Matriz"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "last_login_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "deleted_at": {
                        "description": "Sempre presente; null quando ativo.",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Envelope": {
                "description": "Envelope padrão de toda resposta.",
                "required": [
                    "status",
                    "message",
                    "errors",
                    "data"
                ],
                "properties": {
                    "status": {
                        "description": "Definições globais da OpenAPI (info, servidor, segurança, schemas e respostas reutilizáveis).\nOs endpoints são anotados nos próprios controllers. O atributo não aceita expressão, então\ntítulo e descrição usam os marcadores de OpenApiBuilder::BRAND_TOKENS, preenchidos com a\nmarca do config (app.name / app.full_name) na geração.",
                        "type": "string",
                        "enum": [
                            "success",
                            "error"
                        ]
                    },
                    "message": {
                        "type": "string",
                        "example": "Operação realizada com sucesso."
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "data": {
                        "description": "Objeto, lista ou null conforme o endpoint (qualquer tipo)."
                    },
                    "meta": {
                        "$ref": "#/components/schemas/Pagination"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "Unauthenticated": {
                "description": "Não autenticado (401).",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorEnvelope"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "Acesso negado (403).",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorEnvelope"
                        }
                    }
                }
            },
...
            "ValidationError": {
                "description": "Dados inválidos (422).",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationErrorEnvelope"
                        }
                    }
                }
            },
            "TooManyRequests": {
                "description": "Limite de requisições excedido (429).",
                "headers": {
                    "Retry-After": {
                        "description": "Segundos até poder tentar novamente.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                },
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorEnvelope"
                        }
                    }
                }
            },
...
        },
        "parameters": {
            "Page": {
                "name": "page",
                "in": "query",
                "description": "Página (1..n).",
                "schema": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1
                }
            },
            "PerPage": {
                "name": "per_page",
                "in": "query",
                "description": "Itens por página: máx. 100; até 200 com `is_active=1` (selects).",
                "schema": {
                    "type": "integer",
                    "default": 15,
                    "maximum": 200,
                    "minimum": 1
                }
            },
            "Search": {
                "name": "q",
                "in": "query",
                "description": "Busca textual (contém, sem diferenciar maiúsculas; `%` e `_` são literais).",
                "schema": {
                    "type": "string",
                    "maxLength": 100
                }
            },
            "WithTrashed": {
                "name": "with_trashed",
                "in": "query",
                "description": "Inclui registros excluídos (soft delete). Exige `*.manage` do recurso; sem ela → 403.",
                "schema": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                }
            },
            "IsActive": {
                "name": "is_active",
                "in": "query",
                "description": "Filtra por ativo (1) / inativo (0).",
                "schema": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtido em POST /auth/login (expira em 12 h).",
                "bearerFormat": "Sanctum personal access token",
                "scheme": "bearer"
            }
        }
    },
...
}